vue export, tests + docs
