+++ addons/background/src/index.js
@@ -1,7 +1,66 @@
-import { window } from 'global';
-import ReactBackground from './react';
-import VueBackground from './vue';
+import React from 'react';
+import { polyfill } from 'react-lifecycles-compat';
+import PropTypes from 'prop-types';
 
-const Background = window.STORYBOOK_ENV === 'vue' ? VueBackground : ReactBackground;
+import addons from '@storybook/addons';
 
-export default Background;
+export class BackgroundDecorator extends React.Component {
+  constructor(props) {
+    super(props);
+
+    const { channel } = props;
+
+    // A channel is explicitly passed in for testing
+    if (channel) {
+      this.channel = channel;
+    } else {
+      this.channel = addons.getChannel();
+    }
+
+    this.state = {};
+  }
+
+  componentDidMount() {
+    this.channel.emit('background-set', this.props.backgrounds);
+  }
+
+  componentWillUnmount() {
+    this.channel.emit('background-unset');
+  }
+
+  render() {
+    return this.state.story;
+  }
+}
+
+BackgroundDecorator.getDerivedStateFromProps = ({ story }, { prevStory }) => {
+  if (story !== prevStory) {
+    return {
+      story: story(),
+      prevStory: story,
+    };
+  }
+  return null;
+};
+
+BackgroundDecorator.propTypes = {
+  backgrounds: PropTypes.arrayOf(PropTypes.object),
+  channel: PropTypes.shape({
+    emit: PropTypes.func,
+    on: PropTypes.func,
+    removeListener: PropTypes.func,
+  }),
+  // eslint-disable-next-line react/no-unused-prop-types
+  story: PropTypes.func.isRequired,
+};
+
+BackgroundDecorator.defaultProps = {
+  backgrounds: [],
+  channel: undefined,
+};
+
+polyfill(BackgroundDecorator);
+
+export default backgrounds => story => (
+  <BackgroundDecorator story={story} backgrounds={backgrounds} />
+);
